[settings] Use nicer directory for daemon settings
Use `/etc/xdg` as central system-wide directory for daemon settings,
rather than root home. This directory is more amenable in the
permissions it allows, freeing us from parent-dir checks when the user
forgets sudo. It is the same directory that `QSettings` would use in
`SystemScope`. Drop the `multipass` leaf dir, to avoid having to
check-for/create it in the daemon.